+++ packages/local-prover/scripts/genJoinSplitTestCase.ts
@@ -1,7 +1,8 @@
 import findWorkspaceRoot from "find-yarn-workspace-root";
 import * as path from "path";
 import * as fs from "fs";
-import { poseidonBN } from "@nocturne-xyz/circuit-utils";
+import * as JSON from "bigint-json-serialization";
+import { poseidonBN } from "@nocturne-xyz/crypto-utils";
 import { WasmJoinSplitProver } from "../src/joinsplit";
 import { IncrementalMerkleTree } from "@zk-kit/incremental-merkle-tree";
 import {
@@ -10,7 +11,6 @@ import {
   MerkleProofInput,
   EncodedNote,
   StealthAddressTrait,
-  randomFr,
   encodeEncodedAssetAddrWithSignBitsPI,
   TreeConstants,
 } from "@nocturne-xyz/sdk";
@@ -40,6 +40,10 @@ const vk = nocturneSigner.vk;
 const stealthAddrA = nocturneSigner.canonicalStealthAddress();
 const stealthAddrB = nocturneSigner.canonicalStealthAddress();
 const spendPk = nocturneSigner.spendPk;
+
+const refundAddr = StealthAddressTrait.compress(
+  nocturneSigner.generateRandomStealthAddress()
+);
 
 // Two old notes: 100 + 50 = 150
 const oldNoteA: EncodedNote = {
@@ -145,13 +149,11 @@ const operationDigest = BigInt(12345);
 const opSig = nocturneSigner.sign(operationDigest);
 console.log(opSig);
 
-const encRandomness = randomFr();
-const encSenderCanonAddr = nocturneSigner.encryptCanonAddrToReceiver(
-  nocturneSigner.canonicalAddress(),
-  encRandomness
+const encodedAssetAddrWithSignBits = encodeEncodedAssetAddrWithSignBitsPI(
+  oldNoteA.encodedAssetAddr,
+  refundAddr
 );
-
-console.log("encSenderCanonAddr", encSenderCanonAddr);
+const encodedAssetId = oldNoteA.encodedAssetId;
 
 const joinsplitInputs: JoinSplitInputs = {
   vk,
@@ -166,12 +168,12 @@ const joinsplitInputs: JoinSplitInputs = {
   newNoteB,
   merkleProofA: merkleProofAInput,
   merkleProofB: merkleProofBInput,
-  encRandomness,
-  encodedAssetAddrWithSignBits: encodeEncodedAssetAddrWithSignBitsPI(
-    oldNoteA.encodedAssetAddr,
-    encSenderCanonAddr
-  ),
+  refundAddr,
+
+  encodedAssetAddrWithSignBits,
+  encodedAssetId,
 };
+
 console.log(joinsplitInputs);
 
 (async () => {