+++ packages/sdk/src/primitives/types.ts
@@ -1,16 +1,14 @@
-import {
-  CanonAddress,
-  CompressedStealthAddress,
-  EncryptedCanonAddress,
-} from "../crypto";
-import { CompressedPoint } from "../crypto/pointCompression";
+import { CanonAddress, CompressedStealthAddress } from "../crypto";
 import { JoinSplitInputs, MerkleProofInput, SolidityProof } from "../proof";
-import { Asset, EncodedAsset } from "./asset";
 import { IncludedNote, Note } from "./note";
+import { EncodedAsset } from "./asset";
+import { SerializedHybridCiphertext } from "@nocturne-xyz/crypto-utils";
 
 export const BN254_SCALAR_FIELD_MODULUS =
   21888242871839275222246405745257275088548364400416034343698204186575808495617n;
 export const BLOCK_GAS_LIMIT = 30_000_000n;
+export const SENDER_COMMITMENT_DOMAIN_SEPARATOR =
+  5680996188676417870015190585682285899130949254168256752199352013418366665222n;
 
 export type Address = string;
 
@@ -43,16 +41,10 @@ export interface OperationResult {
   numRefunds: bigint;
 }
 
-export interface EncryptedNote {
-  owner: CompressedStealthAddress;
-  encappedKey: bigint;
-  encryptedNonce: bigint;
-  encryptedValue: bigint;
-}
+export type EncryptedNote = SerializedHybridCiphertext;
 
 export interface IncludedEncryptedNote extends EncryptedNote {
   merkleIndex: number;
-  asset: Asset;
   commitment: bigint;
 }
 
@@ -76,18 +68,20 @@ export interface PreSignJoinSplit extends BaseJoinSplit {
   newNoteB: Note;
   merkleProofA: MerkleProofInput;
   merkleProofB: MerkleProofInput;
+  senderCommitment: bigint;
+  refundAddr: CompressedStealthAddress;
 }
 
 export interface PreProofJoinSplit extends BaseJoinSplit {
   opDigest: bigint;
   proofInputs: JoinSplitInputs;
-  encSenderCanonAddr: EncryptedCanonAddress;
+  senderCommitment: bigint;
+  refundAddr: CompressedStealthAddress;
 }
 
 export interface ProvenJoinSplit extends BaseJoinSplit {
   proof: SolidityProof;
-  encSenderCanonAddrC1: CompressedPoint;
-  encSenderCanonAddrC2: CompressedPoint;
+  senderCommitment: bigint;
 }
 
 interface BaseOperation {
@@ -161,12 +155,12 @@ export interface OptimisticOpDigestRecord {
   metadata?: OperationMetadata;
 }
 
-export type ActionMetadata = {
+export interface ActionMetadata {
   type: "Transfer";
   recipientAddress: Address;
   erc20Address: Address;
   amount: bigint;
-};
+}
 
 export interface OperationMetadata {
   action: ActionMetadata;