+++ packages/sdk/test/closableAsyncIterator.test.ts
@@ -3,7 +3,7 @@ import chai, { expect } from "chai";
 import sinon from "sinon";
 import sinonChai from "sinon-chai";
 import { ClosableAsyncIterator } from "../src";
-import { randomBigInt } from "../src";
+import { randomBigInt } from "./utils";
 
 chai.use(sinonChai);