+++ packages/circuits/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- force `refundAddr` to be owned by the sender
+- replace ElGamal encryption with `senderCommitment`
 - add `IsOrderGreaterThan8` template and use it to check BJJ points given as input in joinsplit circuit
 - skip contributions for subtree update circuit in dev
 - clean before building C++ witness generators